requirements.txt
streamlit
absl-py
sacrebleu
pandas
streamlit-toggle-switch
rouge_score
torch==2.0.0
transformers==4.28.1
huggingface-hub==0.13.4
indobenchmark-toolkit==0.1.6
datasets==2.11.0
sklearn-pandas==2.2.0
nltk==3.8.1
